pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fluxion"
version = "0.2.0"
description = "Fluxion — GPU-accelerated math & physics visualization library"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name="Ahmed Isa Zaweel", email="ahmedzaweel@gmail.com"}]

dependencies = [
  "moderngl>=5.6",
  "pyrr>=0.10",
  "numpy>=1.23",
  "pillow>=9.0",
  "imageio>=2.25",
  "imageio-ffmpeg>=0.4",
  "matplotlib>=3.5"
]
